Move telemetry opaques out of OpaqueVal.h
